move fixture to `test_analyses.py`
